+++ packages/react/src/I18nProvider.tsx
@@ -1,9 +1,10 @@
-import React, { FunctionComponent, ComponentType, ReactNode } from "react"
+import React, { ComponentType, FunctionComponent } from "react"
 import { I18n } from "@lingui/core"
+import { TransRenderProps } from "./Trans"
 
 export type I18nContext = {
   i18n: I18n
-  defaultComponent?: ComponentType<{ children?: ReactNode }>
+  defaultComponent?: ComponentType<TransRenderProps>
 }
 
 export type withI18nProps = {
@@ -12,6 +13,7 @@ export type withI18nProps = {
 
 export type I18nProviderProps = I18nContext & {
   forceRenderOnLocaleChange?: boolean
+  children?: React.ReactNode
 }
 
 const LinguiContext = React.createContext<I18nContext>(null)